Создайте новую модель блоговой записи и новое приложение blog.
